interpret column header suffixes and allow custom colors
Here we use microReact-style interpretation of column header names. Specifically, a double-underscore in a header name is interpreted as extra information, with valid suffixes "colour", "autocolour" and "shape".

We ignore "shape" (as auspice doesn't have this capability) and "autocolour" (as that's our default).

We parse "colour" as a column defining colours for the corresponding column. We enforce that these are (long) HEX values, but this should be relaxed in the future. Where multiple nodes with the same trait value define different colours, we average them similarly to the map.