Remap key "d" to open download modal
Previously this triggered an automatic screenshot (SVG) download, which was dev-only, undocumented functionality that made it into production.

Here we map "d" to open the download modal which is both more expected, and helpful for the current work. This key shortcut can either be removed in this PR or documented and released as-is.