Allow tip labels to be user selected
Adds an additional dropdown box to allow users to change the attribute used as tip labels. This is similar in design to the branch label drop down.

Currently attributes are limited to the colorings. The function which collects these, `collectAvailableTipLabelOptions` could be easily extended to produce further valid attributes.

This should pave the way for smarter and more customisable logic around _when_ labels are displayed, and the dropdown should not be shown in the situation that labels aren't shown.

For a full description, see #1201
Closes #1201