Update filters to hide ancestral nodes before last common ancestor
Before this commit, filters determined visibility by:
1. Finding all terminal nodes in view window that match the filters and
mark them as visible.
2. Recursively marking all parent nodes of these terminal nodes visible,
all the way up to the root of the tree.

This commit adds an additional step to hide ancestral nodes before
the last common ancestor of visible terminal nodes. Starting from
the root of the tree, recursively hide each node if it does not have
more than one child node that is visible.