[bugfix] allow legend filtering to work with genotype colorings
The improvements to legend-item filtering introduced in PR #1150 treated genotype like a "normal" coloring, however it is not (currently) possible to get the genotype of a node via `getTraitFromNode`. This resulted in empty legends for genotypes. This commit fixes this by adding a check within the `createVisibleLegendValues` function. A future improvement may be to extend `getTraitFromNode` to be able to return the genotype when appropriate.