Update hideNodesAboveVisibleCommonAncestor() to return index of LCA
In preparation for adding the zoom-to button, return the index of the
node that represents the last common ancestor of the filtered nodes.

If we can't find the last common ancestor for some reason (e.g. the LCA
is not within the current view), the function returns null as default.