Move animation controls to the sidebar
The play / pause / reset buttons were originally placed in the upper left corner of the map to highlight this functionality for Zika and Ebola where animation works well. With emphasis on SARS-CoV-2, we should no longer highlight animation. It doesn't work well for the large SARS-CoV-2 trees.

Placing play / pause / reset under the date slider also helps to show what's actually going on with the animation.