Update webpack chunk-hash-generation settings
This removes `HashedModuleIdsPlugin` from the build pipeline as
we use `WebpackChunkHash` to generate contenthash for each chunk.

This caused hard-to-debug bugs in auspice (and nextstrain.org)
where a hash was being generated from the pre-minified (etc)
bundle contents, but different minification resulted in different
bundle contents. As the bundle name was the same, clients could
use their (out-of-date) cached version, which would result in
a variety of errors in auspice.