Collect all variable states observed in the tree
Refactor code to expose a function which collects all genotype states in the tree with the exception of non-variable sites (for which the option root-sequence JSON would be needed).

This updates the behavior to preserve filter-by-genotype URL queries on load which define basal states for variable sites. If the site is not variable, then the filter is removed as it makes little sense to filter to non-variable sites.

Note that the filtering algorithm needs to be updated (future commit) to take into account basal states.