Genotype filters at same residue operate as OR
Previously genotypes at the same residue (i.e. same gene and position)
would intersect ("AND"), and thus couldn't be used. We now treat these
as unions ("OR"). This allows useful queries such as
`?gt=HA1.186D,186S`

Note: The page info header does not accurately express the behavior of
genotype filters. This will be improved in a subsequent commit.