Find MRCA when using genotype filters
Previously we were not recomputing the MRCA of the the filtered nodes if genotype filters were applied, which resulted in the "zoom to selected" button behaving as if the genotype filters did not exist.

For non-genotype ("normal") filters, given a set `X` of visible nodes, we simultaneously find the MRCA  of `X` and add the paths from the MRCA to `X`.

For genotype filtering, we wish to find the MRCA of `X` _without_ modifying `X`.  Note that this allows situations where `MRCA \not\in X`. This is introduced in this commit via the function `findFilteredMRCA` which uses a 3 step process:

1. Find the basal-most nodes of each (potentially non-monophyletic) visible cluster
2. Identify the paths from the root to the nodes in (1).
3. Find the first fork in this set of paths