Make dataset loading more consistent
These changes were motivated by
[#1283](https://github.com/nextstrain/auspice/issues/1283) which arose
as we used different code paths for loading a dataset viz and a
narrative.

Here we represent each dataset by a `Dataset` object. This is used for
stand alone datasets, each dataset in a tangletree, and each dataset in
a narrative. Each dataset instance describes the various API endpoints
of datafiles for each dataset, manages fetching of these datafiles and,
where appropriate, can dispatch data to update redux state.

This has been tested on various single datasets, tangle-trees, and
narratives in this repo. Notably, this commit breaks narratives with
multiple datasets; this will be fixed in a subsequent commit to
reflect Eli's work in PR #1312.