narratives load freqs & root seq from cache;
As James and I sketched out verbally
over the phone, this tries to set up
then functions for promises in the dataset
cache which are initiated at the start
of loading a narrative with sidecar files.
This has been minimally tested with
the frequencies file and seems to work
on the surface.