Add support for "data_provenance" metadata
In the early stages of COVID-19, we added support for acknowledging
GISAID as the source of data in the Byline. This was inferred based
on domain / dataset name heuristics.

We now support data provenance to be defined in the dataset JSON
(see https://github.com/nextstrain/augur/pull/705) and all core
nCoV builds have been updated to include this here. This commit
parses and renders such information.

Note that a previous commit removed the "Build info" from the byline
for datasets displaying GISAID (see [0]), which I believe was an
oversight. This commit reinstates it.

[0] https://github.com/nextstrain/auspice/commit/18d5d21ac731d771c97bbefdc6ca20854fd1e2c2