[PhyloTree] branch label bugfix
Fixes a bug where we sometimes ask PhyloTree to update the branch labels
for a view without any branch labels, which would cause auspice
to crash.
I first tried to fix this in a80e1869956daa0c63a0bfc5f564745023263358
but that didn't cover all the situations when this could arise.