Legend bound matching is (a, b] for continuous scales
This restores the algorithm used to associate a hovered legend
item to tips for continuous variables. Commit
0f37b1ac2bff354115e47e14719e7fef4cb282fb (Mar 2018) incorrectly changed
this to `tip \in [a, b]` rather than the intended (and documented)
`tip \in (a, b]`.

This takes on more importance given that the previous commit allows
user-defined bounds.

Note that the frequencies panel already used `(a, b]` matching, so now
the legend matching mirrors this.