Inactivate tips when no longer focused
Selecting a tip (and the corresponding info panel modal) is intrinsically linked to setting a single active "strain" filter. When clicking away from this modal the filter was removed.

This caused issues in the situation where "a tip is selected under a filter, then you click on that tip to see more information, when you click away to close the box, the filter is removed."

The solution here is to inactivate the strain filters when removing the info panel. This has the benefit (bug?) of keeping a history of clicked tips in the form of inactive filters. (Note that all strain filters can be removed by a single click in the sidebar filter badge.)

Closes https://github.com/nextstrain/auspice/issues/1357