Only publish to npm using Node.js 14
Our previous config would run `npm publish` for each node_js
version (previously this was n=2, it's currently n=3). This
resulted in errors as we were trying to publish the same (auspice)
version number multiple times. This may or may not be related to
the integrity shasum errors for auspice 2.29.0.