drag-and-drop: support metadata in Excel format
Uses the xlsx package to convert dropped files to a CSV string that
can then be parsed by Papa Parse. Currently only reads the first sheet
of Excel files.

We continue to use Papa Parse instead of using `xlsx.utils.sheet_to_json`
because papaparse has built in options for `comments` and `dynamicTyping` that
are not available in xlsx.