[tree] improve x axis padding
The clipping introduced in v2.30.0 often obscured the stem of the root node for rectangular trees, resulting in it being hard to click on this branch to zoom out. Here we modify the domain of the x axis to allow space for such a branch to be rendered, iff the tree is zoomed.

Note that this does not solve the bigger issue of hard-to-find root branches!