[tree] add button to zoom out
Our previous method for zooming out worked well when the root-branch of the clade itself had some horizontal length (i.e. it wasn't a very short temporal branch, or a branch with ~zero divergence). This proved problematic in many situations, especially when viewing outbreak trees which often had clades with little divergence.

The solution here adds a small icon button to the current row of buttons above the tree. This icon becomes active when the tree is zoomed.

Note that this is only implemented for the main (left) tree at the current time. A second (tangle) tree can still be zoomed by the previous method.