[tree] zoom out of both trees
The zoom out button also zooms out of the second tree if its root index
is not 0. This means the zoom out button will affect both trees if they
are both zoomed in. If the user wants to zoom out of one specific tree,
they would have to use the old method of clicking on the root branch
of the tree.

This pattern makes sense to me since the user already has to click on
each tree individually to zoom in. The previously existing buttons for
zoom to selected and reset layout also affects both trees.