[tree] button to go back to previous zoom
The ability to return to the previous zoom level is useful when
exploring trees in detail, especially when combined with the
magnifying glass (which zooms out).

A hover info box has been added to convey the functionality of this
icon.

The implementation involving multiple trees is conceptually more
complicated, as I don't want us to return to the previous zoom state
in each tree. (This is similar to the magnifiying glass icon, which
zooms out in both trees.) The best solution is probably tree-specific
UI. For now, this functionality is only surfaced if we are viewing a
single tree.