Remove postinstall step
The postinstall script created `./data` and `./narratives` directories, but is now obsolete. The `get-data` script will create its own `./data` directory, and the docs identify when this should be run. There is no script to automatically fetch narratives.

Removing this from the postinstall step means that `env` is not necessary for an auspice install, which should facilitate installation on windows.