Limit scope of narratives search in URL
Changes the substring to search a given URL for narratives content to
include a leading and trailing forward slash. This change limits the
scope of the search in the URL to match paths like
https://nextstrain.org/narratives/ and
https://nextstrain.org/community/narratives/ but to avoid trying to load
as narratives datasets that happen to have the word "narratives" as a
substring of their names (e.g.,
"nextstrain-narratives-demo_ncov_2020-03-04.json").