Update variable names
Variable names changed to better convey that these values represent node order - in rectangular layouts these are the same as y positions, but not for other layouts. Storing these on the <phyloNode> is consistent with other layout (position) variables.

We now have:
<phyloNode>.displayOrder
                        .displayOrderRange
                        .y: y position in domain. Depends on layout.
                        .py: y position of parent node.
                        .yTip: y position in pixels (i.e. the range)
                        .yBase: yTip of parent node.

Note that the untangling code is not currently used, but has been tested here by turning on `globals.attemptUntangle`.