Fix (new) linting errors
These weren't flagged as errors previously, however the previous commit
swapped the rule "no-use-before-define" to "@typescript/no-use-before-define".

These seem to be false-positives (the code works just fine in practice)
related to named exports. For instance, the following "works"
but fails linting:
```
export const a = () => b; // error: 'b' was used before it was defined.
export const b = 42;
```
This is an error even if we enable the "allowNamedExports" option - see
https://eslint.org/docs/latest/rules/no-use-before-define#allownamedexports

I don't know enough about specifications to know why this is considered
an error but still works in practice, however it's not much effort to
reorganise the code in this file.