Categorise mutations in tip-clicked panel
These changes were motivated by issue #1444 [1] where separating
mutations into categories can aid both QC and biological interpretation.
I chose to use "mutations" to refer to mutations observed on a branch
and "changes" to refer to the collection of mutations between a tip
and the root.

The categories are not necessarily disjoint, as a mutation back
to the root will also be a homoplasy or a unique mutation.

Note that changes between a tip sequence and the root aren't grouped
into homoplasies, as a single change (A→C) may be the result of multiple
mutations (e.g. A→B→C) and thus we would need to check the tip state of
each position which is difficult with the current code.

On-hover panels are left unchanged in this commit.

[1] https://github.com/nextstrain/auspice/issues/1444