Tip hover shows counts of changes to root
Previously we displayed the mutations on the branch leading to the tip,
which was inconsistent with the information shown when clicking on a
tip. Showing a summary of counts is a good compromise between visual
space and conveying all the information.

Note that hovering on the branch leading to the tip still shows the
mutations leading to the tip, if that info is desired.