Fetch and load measurements sidecar JSON file
The measurements sidecar JSON file should have a "_measurements" suffix
in the filename. The JSON is expected to have a "collections" array that
contains individual collections with their respective display configs
and measurements data. See more details about JSON schema within the
augur repo.

On loading the data, an id and jitter value are added to each
measurement. The id is just the measurement's array index but is added
to ensure a stable id even if the list of measurements has been filtered.
The jitter will be used as the y-value for each displayed raw
measurement so they are spread apart vertically within each subplot.

The collections array and default collection to display are added to the
measurements redux store. The display config for the default collection
object is parsed and added to the controls redux store.