Measurements: color SVG with tree strain colors
Updates the custom React Redux selector to parse both tree strain
visibility and colors so that it only has to loop through the tree nodes
once.

The `useEffect` hook for coloring the SVG is dependent on `svgData`
to color the SVG every time it gets redrawn and it's dependent on
`treeStrainColors` to color the SVG every time the tree colors change.