[explode] recalculate branch thickness etc
We default to zooming out completely whenever we explode the tree. There are nicer behaviours here, such as re-calculating the MRCA of visible nodes, but this comes at the cost of increased complexity.

Recalculating tip counts is needed for (a) branch thicknesses and (b) forthcoming work which will hide subtrees with no tips.