Measurements: add field values to "Filter Data"
Create a Map of fields with their display titles and unique values for
filter fields for each collection during the loading of the measurements
JSON. This is then used to add measurements options for "Filter Data".
A custom internal `_type` key is added to the options so that they can
be differentiated from the tree filters for dispatching separate actions.
We may want to update our `react-select` package to use their "group"
functionality to visually separate tree and measurement filter options
within the dropdown menu.

The measurements are now filtered by both visible strains in the tree
and the active measurements filters. If there are filters for the
grouping field, then the groups' subplots are displayed in the order
that the filters were added. This gives the users the ability to custom
order the subplots without having to add drag and drop.