[explode] Separate radial subtrees
Note that in radial view separation is expressed by increasing the
angle. If the root of a subtree is near the start (div or time) then
changing the angle doesn't convey a sense of separation. One way to
improve this would be to add a inner circle which the innermost nodes
can then be spaced around, however this wouldn't be appropriate for
non-exploded trees. Worth revisiting in the future, perhaps.