Update all uses of `Select` from `react-select`
With the complete rewrite of `react-select` from v1 to v2, there are a
couple changes to the props of the `Select` component that affect us:

1. `clearable`, `multi`, `searchable` props have been renamed to
`isClearable`, `isMulti`, `isSearchable`
2. `valueKey` has been removed so we must provide `getOptionValue` for
potential Symbol values
3. `value` no longer accepts a simple string. It now always expects
an array of options or a single options object.

The following commit will address the `Select` component from
`react-virtualized-select` used for the data filter.