Measurements: set hover handlers during drawing
Hover handlers were originally set in a separate effect hook that ran
after the SVG was redrawn. This was done to separate out concerns
because I did not realize that CSS selection can be a potential
bottleneck.

This change memoizes the `handleHover` function so that it can be used
as a stable dependency for the drawing functions of the raw measurements
and the means.