Measurements: add padding around overall mean
Prevent color-by means from overlapping with the overall mean by
shifting the yValue of the color-by means if they are too close to the
overall mean.

This does mean that there is a "gap" in middle of each subplot if the
overall mean display is toggled off. I think this is acceptable because
the yValues have no meaning in the subplot. We don't want to do extra
redraw of the color-by means when the overall mean is toggled on/off.