Add fallback for Measurements panel
Use the `PanelSpinner` as a fallback component for Measurements panel
and remove condition on `this.props.measurementsLoaded`. This ensures
that the panel is included on page load so that there is no
interruption in the page layout when the measurements data finishes
loading.

Now the content of the panel is dependent on `measurementsLoaded` and
will only display the MeasurementsPlot when the data has loaded.