Measurements: make jitter randomly distributed
This allows the raw measurement plots to spread over the entire subplot
instead of an even distribution along the y-axis.

The yScale is no longer the extent of the measurement's jitter values
because the min and max are now predefined. The measurement's diameter
is also accounted for in the domain so that they do not get cut off when
they are plotted on the edge of the subplot.