Measurements: use stable order for subplots
The default ordering for subplots is based on the total count of
measurements per grouping value, with the highest count first.
The default order can also be customized within the measurements JSON.
The JSON defined order is given precendence, but any values not included
in the order will follow the default ordering by measurements count.

The order can still be defined by the user using the filters UI.