Measurements: fix filter ordering of subplots
The ordering of filtered subplots was broken by 5ae8f23 because the
provided filter order was no longer a Map, but a flat Array of values.

This commit changes the function to only expect a single flat array of
values instead of two arrays. This moves the definition of groupings'
order out of `groupMeasurements` to the Measurements component so that
the determination of the order of subplots is not hidden as a side
effect in the grouping function.