[tree] escape key dismisses any info modal
This implements what we think is an intuitive UI, and also helps with
the bug described in #1243 where removing/disabling a strain-filter
results in the tree visualisation falling out-of-sync with the modal.

Note that `event.key` may not be implemented in all browsers, however
this functionality is a nicety rather than a must-have, so this is
tolerable. Thanks to @joverlee521 for suggesting to use this instead of
`event.keyCode` during PR review.

See https://github.com/nextstrain/auspice/issues/1479 for context.

Closes #1479