Drag-n-drop metadata shows in second tree
Tips present only in the second tree will be correctly coloured, and
appear in the legend. However there are two (long-standing) bugs /
shortcomings related to the second tree which are relevant here:

1. Trait values which only appear in tips in the second tree will not be
available as filtering options (they will be correctly shown on the tree
and in the legend, however).

2. The tip-counts, as they appear in the header, only consider the main
(left) tree. For instance, if you have a metadata CSV with information
on 100 tips in the left tree, and 150 tips in the right tree (and 100
are in both), and you filter to the strains which appear in the dropped
CSV (the colouring is the same name as the file). The header will now
indicate "Showing 100 of XXX genomes", despite the fact that all 150
strains in the second tree are displayed.

Both of these bugs are not specific to drag-n-drop metadata, the same
situations happen with any colouring / filter.

Closes #1488