Fix: narrative dataset filters
When users change to a slide in a narrative that has the same dataset as
the previous slide, `restoreQueryableStateToDefaults` resets controls
to the _app_ default state which updates the filters state to an empty
object. This leads to the bug described in #1492.

Calling `modifyStateViaMetadata` after `restoreQueryableStateToDefaults`
ensures that all _dataset_ default controls are restored as well.

Resolves #1492.