add getDarkerColor() to colorHelpers
add getDarkerColor() to as inverse to getBrighterColor(); set legend item fill and tree tip fill colors to be the same (returned from the color map), and make the stroke of each relative via getDarkerColor() rather than setting the fill via getBrighterColor(). This helps ensure the fill color of the two is the same and not subject to mismatches due to imprecision in calculating Math.pow(base,k) from D3 brighter/darker.

remove unused getBrighterColor()