Detect duplicate or missing node names
Every now and again we have an issue with a dataset with duplicate
node names, which results in an incredibly uninformative auspice
stack trace. While this is ultimately a problem with the dataset itself
it's simple to detect this and change it to a unique name within
Auspice (with a helpful console message). For good measure we also
deal with the case where a node name was missing entirely.

Note that both of these datasets would fail `augur validate export-v2 <JSON>`.