Allow date filtering for partially temporal trees
Previously a tree was considered temporal (i.e. you could use a temporal
tree metric) or not based on the presence of `num_date` on the root
node. The allowed partially temporal trees (e.g. those where some tip
dates were known) to be displayed as div trees with a colouring of
`num_date`, however the date slider wasn't available.

Here we allow "partially temporal trees" which are defined as tree(s)
where the number of nodes with `num_date` is more than zero but fewer
than the total node count. These get a date-range filter, but aren't
able to be displayed as true time trees [1].

Note that this has one strange / misleading effect. For tips without
a `num_date` they are never filtered out by the date slider. This
is obvious when colouring by `num_date` (they are grey) but when using
a different colouring one must consider that a node may be visible
because it is inside the selected date range _or_ it doesn't have a
date!

[1] I explored hiding nodes without date information but the results
weren't ideal - it could be done, but it would require modifications to
the tree layout code.