fix: prevent crash on unexpected gisaid_epi_isl
In cases where the `gisaid_epi_isl` value does not fit the expected
format with two underscores ("EPI_ISL_<number>"), Auspice crashes when
you click on a tip with an error:

TypeError: Cannot read properties of undefined (reading 'length')

This commit adds a check to verify the `gisaid_epi_isl` value matches
our expected format with a regex (`EPI_ISL_[0-9]{5,}`) before trying to
parse the value to create the acknowledgement URL. The regex specifies
that the accession number must be at least 5 digits so that it can be
sliced to create the proper URL, which requires the last 4 digits of the
accession number:

EPI_ISL_406798 -> https://www.epicov.org/acknowledgement/67/98/EPI_ISL_406798.json

Values that do not match the regex are given the default GISAID URL
("https://gisaid.org"). This strict check only affects the generated
link for the accession number when a user clicks on a tip in the tree.