measurements: truncate long titles
Previously, long titles will push the measurements panel into a separate
line when viewing in "grid" mode¹. This commit updates the CSS for the
title of the Measurements panel so that long titles are truncated with
an ellipsis and do not interrupt the layout of the page.

I had considered another option to wrap the long title instead of
truncating it, but this pushes the panel slightly out of line with the
tree panel.

¹ https://github.com/nextstrain/auspice/pull/1452#discussion_r814392097