measurements: extract handleHover from SVG drawing
This commit decouples the handling of hover panels in mouseover/mouseout
events from the SVG drawing functions. This is done in anticipation that
we will be adding the color-by attribute to the hover panel title. With
the color-by attribute titles, the handleHover function will be
dependent on color-by changes, which we do _not_ want to force SVG
redrawing.