measurements: Add color-by attr as hover panel title
For both raw measurements and the color-by means, we use the color-by
attribute as the title of the hover panel. This will make it easier to
parse what the color means for each element in cases where the legend
contains too many colors.

Note the overall means' hover panels do not have a title because they
are not linked to the color-by attribute.