Separate out undeletion mutations on branches
Undeletions are inferred mutations from a gap to a base, e.g. `-400A`.
In most cases these are bioinformatics errors, so this commit collects
them into a separate category to enhance interoperability of lots of
mutations. The actual base is visible if you click to copy the mutation
list to the clipboard, but in my testing on nCoV datasets this is
always the reference base.