Tests + comments for categorisation of tip changes
Work prompted by the preceeding commit, but I don't think any changes
are actually needed. Note that depending on how the dataset was
generated you can get what look like undeletions when the root
sequence is a gap itself. I think this is rarer than undeletions
on branches so not worth categorising at the present time.
Added tests as the clearest way to document the behaviour here.