measurements: force y-axis labels to scale to fit
In cases where the y-axis grouping label is too long, the text is scaled
down to fit the available space. This does mean that if the text is
extremely long, it can become unreadable. We can improve on this by
manually splitting the text into parts that can fit on multiple lines,
but there's always limits of the available space so punting that for now.