Bugfix when tree zooms back too far
For trees viewed with the divergence metric and with mutations defined,
it is sensible (and easy) for the zoom back button ("-" button) to
walk up the tree until a branch with mutations is found. This will
behave nicely with the common scenario of polytomies encoded by
a bifurcating tree (which often happens during coalescent reconstruction
in augur refine).

The original heuristic remains for those trees where mutations are not
defined on the branches. There are cleverer ways to infer the correct
threshold here if we want to, but I don't think it's worth it at the
moment.

Closes #1529