measurements: increase y-axis label padding
Give more room for y-axis labels to be able to display long reference
strain names without scaling down the text so much that it becomes
unreadable.

This is quick improvement on the y-axis labels after an in person
discussion with @huddlej, but we should definitely dig into better
solutions for displaying long labels. This can be looking more into
SVG text wrapping or even truncating the labels and display the full
text on click modals.