Fix overlappping legend entries
Fixes the bug described in #1547 caused by identical legend entries
which therefore had the same react key and thus DOM reconciliation
failed. The cause of the bug was identified by @joverlee521 in
https://github.com/nextstrain/auspice/issues/1547#issuecomment-1251461561

There is a deeper problem which is why we have
identical entries in the legend, which I'll attempt to solve in the
subsequent commits. However since we deal with user-provided data
it's a good idea to handle the case of duplicated values should it ever
arise.

The array of visible legend entries (array) may change, and thus the
react key shouldn't be reliant on the index. However for duplicate
legend entries there is no other unique identifier available! Combining
the legend value with the index should form a unique key - when legend
values are filtered out the new keys will be different.

Closes #1547