MVP of the narratives debugger
This implements new functionality in auspice available at /edit/narratives
where a user can drag&drop a markdown narrative file & see an overview
of the slides and the datasets they request, as well as preview the
narrative.

This represents a release-ready MVP of the functionality and opens
the door to a number of possible incremental improvements with the
eventual aim of writing and releasing a narrative entirely within-
browser.

This includes changes motivated by review comments from @tsibley
and conversations with @joverlee521 and @j23414