helperFunctions/treeToNewick: use `idxOfFilteredRoot`
In cases where the user has filtered the tree, the `idxOfFilteredRoot`
is the root of the selected tips instead of `idxOfInViewRootNode`.
By using `idxOfFilteredRoot` as the root of the tree, this now acts the
same way as the manual work-around of clicking "Zoom to selected" then
downloading the tree.

However, this does _not_ fix the case where the filter may have created
multiple subtrees. If there are multiple subtrees without a most recent
common ancestor, this function will still produce an empty Newick tree.