measurements: fix disappearing mean/SD on entropy click
Previously, when clicking on an entropy rect, the colored mean/SD
in the measurements panel would flash and disappear. I tracked this down
to the following:
1. Clicking on an entropy rect would trigger two NEW_COLOR dispatches.
2. The second NEW_COLOR action does _not_ update the colors, but the
react-redux `useSelector` hook does a strict reference comparison by
default, so it thinks the `legendValues` (an array) was updated.
3. This would trigger the side-effect `drawMeansForColorBy` but _not_
the `changeMeasurementsDisplay` so the colored mean/SD would get hidden.

This commit updates the `legendValues` to use `lodash.isEqual` equality
function to ensure that we don't run any unnecessary updates. It also
updates the dependencies for the `changeMeasurementsDisplay` side-effect
to make sure they match the dependencies for the `drawMeansForColorBy`
side-effect.